requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.60
langchain==0.3.7
langchain-openai==0.2.14
langchain-google-genai==2.0.7
gdown==5.2.0
numpy==1.26.4
torch==2.2.2
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.9.2
orjson==3.10.7